backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
mollie-api-python
python-dotenv